challenges/misc/cypherpunks/backend/requirements.txt
anchorpy==0.21.0
anchorpy-core==0.2.0
anyio==4.10.0
attrs==25.3.0
base58==2.1.1
based58==0.1.1
blinker==1.9.0
borsh-construct==0.1.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
construct==2.10.68
construct-typing==0.5.6
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonalias==0.1.1
jsonify==0.5
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
pyheck==0.1.5
PyNaCl==1.5.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
typing_extensions==4.15.0
websockets==15.0
Werkzeug==3.1.3
anchorpy==0.21.0
anchorpy-core==0.2.0
anyio==4.10.0
attrs==25.3.0
base58==2.1.1
based58==0.1.1
blinker==1.9.0
borsh-construct==0.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
construct==2.10.68
construct-typing==0.5.6
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonalias==0.1.1
jsonify==0.5
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
pyheck==0.1.5
PyNaCl==1.5.0
requests==2.32.5
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0
Werkzeug==3.1.3
